job-list: move job data into its own files
Problem: `struct job` in job_state.h contains `struct list_ctx` from
job-list.h.  `struct list_ctx` in job-list.h contains
`struct job_state_ctx` from `job_state.h`.  This circular dependency
isn't a problem at the moment, but it can become an issue in the
future when future services are added.

In addition, job_state.[ch] files are getting long.

Solution: Move struct job and job create/destroy functions into
their own files.  Also move jobspec parsing and R parsing into the
new file.
